move datatype folders to super folder
