+++ flow/model/qrecord_batch.go
@@ -10,7 +10,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/jackc/pgx/v5/pgtype"
 
-	"github.com/PeerDB-io/peer-flow/geo"
+	"github.com/PeerDB-io/peer-flow/datatypes/geo"
 	"github.com/PeerDB-io/peer-flow/model/qvalue"
 )
