+++ flow/model/qvalue/dwh.go
@@ -5,8 +5,8 @@ import (
 
 	"go.temporal.io/sdk/log"
 
+	"github.com/PeerDB-io/peer-flow/datatypes/numeric"
 	"github.com/PeerDB-io/peer-flow/generated/protos"
-	"github.com/PeerDB-io/peer-flow/model/numeric"
 )
 
 func DetermineNumericSettingForDWH(precision int16, scale int16, dwh protos.DBType) (int16, int16) {
